fix: resolve registration flow issues
1. Fix weight validation to accept decimal values (72.5 kg)
   - Remove .int() constraint from weightSchema in registration.validation.ts
   - Add test case for decimal weight values

2. Fix transition after registration complete
   - Change registration → chat to registration → session_planning
   - Add validation for registration → session_planning transition
   - User now automatically enters planning phase after registration

3. Fix double greeting issue
   - Remove hardcoded greeting from Telegram bot /start command
   - Bot now sends "hi" to LLM to get personalized greeting
   - Single, contextual greeting from LLM instead of double greeting

All tests passing (225/225)

+++ apps/server/src/domain/user/services/registration.validation.ts
@@ -14,7 +14,7 @@ export type ProfileDataKey = keyof ParsedProfileData;
 const ageSchema = z.union([z.number().int().min(10).max(100), z.null()]).transform((v) => v ?? undefined);
 const genderSchema = z.union([z.enum(['male', 'female']), z.null()]).transform((v) => v ?? undefined);
 const heightSchema = z.union([z.number().int().min(120).max(220), z.null()]).transform((v) => v ?? undefined);
-const weightSchema = z.union([z.number().int().min(30).max(200), z.null()]).transform((v) => v ?? undefined);
+const weightSchema = z.union([z.number().min(30).max(200), z.null()]).transform((v) => v ?? undefined);
 const fitnessLevelSchema = z
   .union([z.enum(['beginner', 'intermediate', 'advanced']), z.null()])
   .transform((v) => v ?? undefined);
